tests/wpt/web-platform-tests/css/css-overflow/parsing/text-overflow-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Overflow: parsing text-overflow with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-overflow/#propdef-text-overflow">
<meta name="assert" content="text-overflow supports only the grammar 'clip | ellipsis'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("text-overflow", 'auto');

test_invalid_value("text-overflow", 'clip ellipsis clip');
</script>
</body>
</html>
